+++ CHANGELOG.md
@@ -11,12 +11,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added i18n to excel exporters of combined equipment
 - added difference value to meter tracking report
 - added translation of TARIFF in myems-web
+- added barchart to Dashboard in myems-web
 - 
 ### Changed
+- updated ChildSpacesTable of Dashboard in myems-web
 - 
 ### Fixed
 - updated datetime picker format pattern
 - fixed integrity rate issue of MeterTracking report
+- fixed MultipleLineChart blank screen issue in myems-web
 - 
 ### Removed
 -